requirements.txt
Django==1.8.3
Markdown==2.6.2
MySQL-python==1.2.5
argparse==1.2.1
django-filter==0.10.0
djangorestframework==3.2.1
wsgiref==0.1.2
